Simplify appContext.xml, mv all *Repository into .domain. packages, rm spmContext.xml (FINERACT-1175)
