+++ fineract-provider/src/main/java/org/apache/fineract/spm/domain/SurveyRepository.java
@@ -16,13 +16,22 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.fineract.portfolio.interestratechart.data;
+package org.apache.fineract.spm.domain;
 
-import org.apache.fineract.portfolio.interestratechart.domain.InterestRateChartSlab;
+import java.util.Date;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface InterestRateChartSlabRepository
-        extends JpaRepository<InterestRateChartSlab, Long>, JpaSpecificationExecutor<InterestRateChartSlab> {
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
 
+    @Query("select s from Survey s where :pointInTime between s.validFrom and s.validTo")
+    List<Survey> fetchActiveSurveys(@Param("pointInTime") Date pointInTime);
+
+    @Query("select s from Survey s ")
+    List<Survey> fetchAllSurveys();
+
+    @Query("select s from Survey s where s.key = :key and :pointInTime between s.validFrom and s.validTo")
+    Survey findByKey(@Param("key") String key, @Param("pointInTime") Date pointInTime);
 }
