+++ fineract-provider/src/main/resources/META-INF/spring/appContext.xml
@@ -34,31 +34,7 @@
     <sec:global-method-security
         pre-post-annotations="enabled" />
 
-    <context:component-scan
-        base-package="org.apache.fineract.accounting.*,
-                                          org.apache.fineract.commands.provider.*,
-                                          org.apache.fineract.commands.handler.*,
-                                          org.apache.fineract.commands.service.*,
-                                          org.apache.fineract.commands.*,
-                                          org.apache.fineract.audit.*,
-                                          org.apache.fineract.infrastructure.creditbureau.*,
-                                          org.apache.fineract.infrastructure.*,
-                                          org.apache.fineract.infrastructure.core.security*,
-                                          org.apache.fineract.scheduledjobs.*,
-                                          org.apache.fineract.organisation.*,
-                                          org.apache.fineract.interoperation.*,
-                                          org.apache.fineract.portfolio.loanaccount.*,
-                                          org.apache.fineract.portfolio.savingsaccount.*,
-                                          org.apache.fineract.portfolio.*,
-                                          org.apache.fineract.useradministration.*,
-                                          org.apache.fineract.mix.*,
-                                          org.apache.fineract.notification.*,
-                                          org.apache.fineract.template.*,
-                                          org.apache.fineract.template.service.*,
-                                          org.apache.fineract.useradministration.*,
-                                          org.apache.fineract.batch,
-                                          org.apache.fineract.adhocquery.*,
-                                          org.apache.fineract.infrastructure.campaigns.**">
+    <context:component-scan base-package="org.apache.fineract.**">
         <context:exclude-filter expression="org.springframework.stereotype.Controller"
             type="annotation" />
 
@@ -73,21 +49,7 @@
     <bean id="auditorAware"
         class="org.apache.fineract.infrastructure.core.domain.AuditorAwareImpl" />
     <jpa:auditing auditor-aware-ref="auditorAware" />
-    <jpa:repositories base-package="org.apache.fineract.commands.domain" />
-    <jpa:repositories base-package="org.apache.fineract.infrastructure.*.domain" />
-    <jpa:repositories base-package="org.apache.fineract.accounting.*.domain" />
-    <jpa:repositories base-package="org.apache.fineract.useradministration.domain" />
-    <jpa:repositories base-package="org.apache.fineract.organisation.*.domain" />
-    <jpa:repositories base-package="org.apache.fineract.portfolio.*" />
-    <jpa:repositories base-package="org.apache.fineract.mix.domain" />
-    <jpa:repositories base-package="org.apache.fineract.scheduledjobs.domain" />
-    <jpa:repositories base-package="org.apache.fineract.template.domain" />
-    <jpa:repositories base-package="org.apache.fineract.infrastructure.campaigns.sms.domain" />
-    <jpa:repositories base-package="org.apache.fineract.adhocquery.domain" />
-    <jpa:repositories base-package="org.apache.fineract.notification.domain"/>
-    <jpa:repositories base-package="org.apache.fineract.infrastructure.campaigns.email.domain"/>
-    <jpa:repositories base-package="org.apache.fineract.interoperation.domain"/>
-
+    <jpa:repositories base-package="org.apache.fineract.**.domain" />
 
     <import resource="infrastructure.xml" />
 
@@ -110,6 +72,4 @@
         <property name="targetMethod" value="setStrategyName" />
         <property name="arguments" value="MODE_INHERITABLETHREADLOCAL" />
     </bean>
-
-    <import resource="spmContext.xml"/>
 </beans>
